Combine LoopPopupProvider and call choreo linking into ActivityWrenchPopupProvider. Also fix one bug
