+++ lib/features/context-pad/ContextPadProvider.js
@@ -230,19 +230,20 @@ ContextPadProvider.prototype.getContextPadEntries = function(element) {
   }
 
   // --------------------------------------------------------------------------------------------
-  // change loop type
+  // Provide a popup for the wrench menu of choreo activities.
+  // Can be used to change loop-type or set references for CallChoreos.
   if (is(businessObject, 'bpmn:ChoreographyActivity')) {
     assign(actions, {
-      'loopMarker.create': {
+      'activityWrench': {
         group: 'edit',
         className: 'bpmn-icon-screw-wrench',
-        title: this._translate('Marker'),
+        title: this._translate('Link CallChoreography'),
         action: {
           click: (event, element) => {
             const position = assign(popupMenuPosition(element), {
               cursor: { x: event.x, y: event.y }
             });
-            self._popupMenu.open(element, 'loop-provider', position);
+            self._popupMenu.open(element, 'activity-wrench-provider', position);
           }
         }
       }
@@ -394,25 +395,7 @@ ContextPadProvider.prototype.getContextPadEntries = function(element) {
     });
   }
 
-  // --------------------------------------------------------------------------------------------
-  // Link CallChoreo
-  if (is(businessObject, 'bpmn:CallChoreography')) {
-    assign(actions, {
-      'linkCallChoreo': {
-        group: 'edit',
-        className: 'bpmn-icon-screw-wrench',
-        title: this._translate('Link CallChoreography'),
-        action: {
-          click: (event, element) => {
-            const position = assign(popupMenuPosition(element), {
-              cursor: { x: event.x, y: event.y }
-            });
-            self._popupMenu.open(element, 'link-provider', position);
-          }
-        }
-      }
-    });
-  }
+
 
   return actions;
 };
