+++ lib/features/modeling/cmd/LinkCallChoreoHandler.js
@@ -19,15 +19,18 @@ LinkCallChoreoHandler.prototype.link = function(element, newLink, oldLink) {
     newLink.callChoreographyActivity.push(element);
   }
   // we unlink the the element with the old choreo
-  if (oldLink !== undefined){
+  if (oldLink !== undefined) {
     oldLink.callChoreographyActivity = oldLink.callChoreographyActivity.filter(bo => bo !== element);
   }
   return [element];
 };
 
 LinkCallChoreoHandler.prototype.preExecute = function(context) {
-  if(context.newLink) {
+  if (context.newLink) {
     context.newLink.callChoreographyActivity = context.newLink.callChoreographyActivity || [];
+  }
+  if (context.oldLink) {
+    context.oldLink.callChoreographyActivity = context.oldLink.callChoreographyActivity || [];
   }
   if (context.newLink === context.oldLink) {
     // if the new linked choreo is already linked with this call choreo, we reset the the link.
@@ -41,7 +44,7 @@ LinkCallChoreoHandler.prototype.execute = function(context) {
 };
 
 LinkCallChoreoHandler.prototype.revert = function(context) {
-  if(context.newLink !== undefined){
+  if (context.newLink !== undefined) {
     context.newLink.callChoreographyActivity.pop();
   }
   return this.link(context.element, context.oldLink, context.newLink);
