+++ lib/features/popup-menu/ActivityWrenchPopupProvider.js
@@ -1,33 +1,41 @@
 import inherits from 'inherits';
 import PopupMenuProvider from 'diagram-js/lib/features/popup-menu/PopupMenuProvider';
+import { is } from 'bpmn-js/lib/util/ModelUtil';
 
 /**
- * Provides a popup menu for setting loop types. In the future this could be extended to
- * also contain a replacement menu like the ReplacementMenuProvider in bpmn-js or this could be
- * done in another module.
  *
  * @param popupMenu
  * @param {ChoreoModeling} modeling
+ * @param choreoUtil
  * @constructor
  */
-export default function LoopPopupProvider(popupMenu, modeling) {
+export default function ActivityWrenchPopupProvider(popupMenu, modeling, choreoUtil) {
   this._modeling = modeling;
+  this._choreoUtil = choreoUtil;
   PopupMenuProvider.call(this, popupMenu);
 }
 
-inherits(LoopPopupProvider, PopupMenuProvider);
+inherits(ActivityWrenchPopupProvider, PopupMenuProvider);
 
-LoopPopupProvider.$inject = [
-  'popupMenu',
-  'modeling'
-];
 
-LoopPopupProvider.prototype.getEntries = function(element) {
-  // TODO do checks
+ActivityWrenchPopupProvider.prototype.getEntries = function(element) {
+  if (is(element, 'bpmn:CallChoreography')) {
+    const currentChoreo = this._choreoUtil.currentChoreography();
+    const currentLink = element.businessObject.calledChoreographyRef;
+    const items = this._choreoUtil.choreographies().filter(choreo => choreo.id !== currentChoreo.id).map(choreo => {
+      const name = choreo.name || choreo.id;
+      return {
+        id: ('select-' + choreo.id),
+        label: name,
+        active: currentLink === choreo,
+        action: () => this._modeling.linkCallChoreo(element.businessObject, currentLink, choreo) };
+    });
+    return items;
+  }
   return [];
 };
 
-LoopPopupProvider.prototype.getHeaderEntries = function(element) {
+ActivityWrenchPopupProvider.prototype.getHeaderEntries = function(element) {
   const isParallel = element.businessObject.loopType === 'MultiInstanceParallel';
   const isSequential = element.businessObject.loopType === 'MultiInstanceSequential';
   const isStandard = element.businessObject.loopType === 'Standard';
@@ -73,6 +81,6 @@ LoopPopupProvider.prototype.getHeaderEntries = function(element) {
   ];
 };
 
-LoopPopupProvider.prototype.register = function() {
-  this._popupMenu.registerProvider('loop-provider', this);
+ActivityWrenchPopupProvider.prototype.register = function() {
+  this._popupMenu.registerProvider('activity-wrench-provider', this);
 };
