+++ lib/features/popup-menu/index.js
@@ -1,22 +1,19 @@
 import PopupMenuModule from 'diagram-js/lib/features/popup-menu';
 
-import LoopPopupProvider from './LoopPopupProvider';
 import ParticipantPopupProvider from './ParticipantPopupProvider';
 import ReplaceMenuProvider from './ReplaceMenuProvider';
-import LinkCallChoreoPopupProvider from './LinkCallChoreoPopupProvider';
+import ActivityWrenchPopupProvider from './ActivityWrenchPopupProvider';
 
 export default {
   __depends__: [
     PopupMenuModule,
   ],
   __init__: [
-    'loopPopupProvider',
     'participantPopupProvider',
     'replaceMenuProvider',
-    'linkCallChoreoPopupProvider'
+    'activityWrenchPopupProvider'
   ],
-  loopPopupProvider: ['type', LoopPopupProvider],
   participantPopupProvider: ['type', ParticipantPopupProvider],
   replaceMenuProvider: ['type', ReplaceMenuProvider],
-  linkCallChoreoPopupProvider: ['type', LinkCallChoreoPopupProvider]
+  activityWrenchPopupProvider: ['type', ActivityWrenchPopupProvider]
 };
